+++ arch/x86/idt.h
@@ -1,0 +1,7 @@
+// arch/x86/idt.h
+#ifndef TOYIX_ARCH_X86_IDT_H
+#define TOYIX_ARCH_X86_IDT_H
+
+void idt_init(void);
+
+#endif